Adapt commands to new command method
Also added runIf to added commands which is a funciton passed the same
args as the command event to determine if the command should run, as
well as a default runIf for some preconfigured command types.
