+++ src/lib/commands.js
@@ -1,29 +1,42 @@
 import { Events } from "./commandEvents.js";
-import { ApplicationCommandOptionTypes,
-	 upsertGlobalApplicationCommands,
+import { upsertGlobalApplicationCommands,
 	 ApplicationCommandTypes,
-         v5 } from "../../deps.js";
+	 InteractionTypes } from "../../deps.js";
 
 export function enableCommandsPlugin(bot) {
-    bot.commands = {};
+    bot.commands = [];
     bot.events = {};
-    for (e of Events) {
-	bot.commands[e] = {};
+    for (const e of Events) {
 	bot.events[e] = function(...args) {
-	    for (command of bot.commands[e]) {
-		if (command.actions.isArray()) {
-		    for (action of command.actions) {
-			if (typeof(action) == "function") {
-			    action(...args);
+	    const commands = bot.commands.filter( comm => {
+		return comm.event == e;
+	    })
+	    for (const command of commands) {
+		if (Array.isArray(command.actions)) {
+		    if (command.runIf == undefined ||
+			command.runIf(...args)) {
+			for (const action of command.actions) {
+			    if (typeof(action) == "function") {
+				action(...args);
+			    }
 			}
 		    }
-		}else {
+		}else { 
 		    /* Will soon be replaced by proper logging */
 		    console.log("Should be unreachable")
 		}
 	    }
 	}
     }
+}
+
+function filterApplicationCommand(_bot, interaction) {
+    if (interaction.type == InteractionTypes.ApplicationCommand ) {
+	if (interaction.data.name == this.name) {
+	    return true;
+	}
+    }
+    return false;
 }
 
 const builtinCommandTypes = {
@@ -33,32 +46,53 @@ const builtinCommandTypes = {
     },
     slash: {
 	event: "interactionCreate",
-	type: ApplicationCommandTypes.ChatInput
+	type: ApplicationCommandTypes.ChatInput,
+	runIf: filterApplicationCommand
     },
     message: {
 	event: "interactionCreate",
-	type: ApplicationCommandTypes.Message
+	type: ApplicationCommandTypes.Message,
+	runIf: filterApplicationCommand
     },
     user: {
 	event: "interactionCreate",
-	type: ApplicationCommandTypes.User
+	type: ApplicationCommandTypes.User,
+	runIf: filterApplicationCommand
     }
 };
 
 export function addBotCommand(bot, command) {
+    const com = command;
     if (com.type in builtinCommandTypes) {
-	for (prop in builtinCommandTypes[com.type]) {
-	    com[prop] = builtinCommandTypes[com.type][prop];
+	const commandType = builtinCommandTypes[com.type]
+	for (const prop in commandType) {
+	    com[prop] = commandType[prop];
 	}
+    }else if (!com.type) {
+	com.type="Undefined";
     }
     if (!com.event) {
 	return null
     }
+    if (com.type != ApplicationCommandTypes.ChatInput) {
+	delete com.options
+	delete com.description
+	delete com.descriptionLocalizations
+    }
     // Command format validated, generate UUID for command
-    com.uuid = v5.generate()
+    com.uuid = crypto.randomUUID()
+    bot.commands.push(com)
     return com
 }
 
 export function updateBotCommands(bot) {
-    upsertGlobalApplicationCommands(bot, bot.commands);
+    const commands = bot.commands.filter(comm => {
+	const upsertedTypes = [
+	    ApplicationCommandTypes.ChatInput,
+	    ApplicationCommandTypes.Message,
+	    ApplicationCommandTypes.User
+	];
+	return upsertedTypes.includes(comm.type);
+    });
+    upsertGlobalApplicationCommands(bot, commands);
 }
